modular: split up graphql schema definitions and resolvers to different files
